+++ modules/Student_Billing/MassAssignPayments.php
@@ -26,6 +26,8 @@ if ( $_REQUEST['modfunc'] === 'save' )
 							'PAYMENT_DATE' => $date,
 							'AMOUNT' => preg_replace( '/[^0-9.-]/', '', $_REQUEST['amount'] ),
 							'COMMENTS' => $_REQUEST['comments'],
+							// @since 11.2 Add CREATED_BY column to billing_fees & billing_payments tables
+							'CREATED_BY' => DBEscapeString( User( 'NAME' ) ),
 						]
 					);
 				}